refactor(shared): extract renter_documents to its own schema module (#459)
Keeps schema.ts at the 800-line cap after the #457/#461/#463 + #459 table
additions piled in (trunk was at 798). Re-exported from schema.ts so drizzle-kit
still discovers it; the users FK uses lazy thunks so the circular import resolves.
